mode thread into its own mod
