+++ sentry/Cargo.toml
@@ -34,7 +34,7 @@ debug-logs = ["log_", "sentry-core/debug-logs"]
 transport = ["reqwest", "native-tls"]
 reqwest = ["reqwest_", "httpdate"]
 curl = ["curl_", "httpdate", "serde_json"]
-surf = ["surf_", "httpdate", "futures"]
+surf = ["surf_", "httpdate"]
 native-tls = ["reqwest_/default-tls"]
 rustls = ["reqwest_/rustls-tls"]
 
@@ -51,7 +51,6 @@ log_ = { package = "log", version = "0.4.8", optional = true, features = ["std"]
 reqwest_ = { package = "reqwest", version = "0.11", optional = true, features = ["blocking", "json"], default-features = false }
 curl_ = { package = "curl", version = "0.4.25", optional = true }
 surf_ = { package = "surf", version = "2.0.0", optional = true }
-futures = { version = "0.3", optional = true }
 httpdate = { version = "0.3.2", optional = true }
 serde_json = { version = "1.0.48", optional = true }
 tokio = { version = "1.0", features = ["rt"] }
